requirements.txt
# PDF
git+https://github.com/ChrizH/pdfstructure.git  # instead of .git @branch_name work as well
sortedcontainers==2.2.2  # Required for pdfstructure
pdfminer.six==20200517  # Required for pdfstructure
pdfplumber~=0.9.0  # Testing purpose

# Basics
streamlit==1.20.0
Pillow~=9.4.0
pypdf2~=2.0.0

future~=0.18.3
fitz~=0.0.1.dev2